# FishingProjectISM/target/classes/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Fishing API
  description : Fishing Project API by Gunce Demirel Ozer
  version: 1.0.0
servers:
  - url: https://example.org/api/v3
paths:
  /user:
    post: 
      tags:
        - user
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createUser
      responses:
        default:
          description: succesful operation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserData'
        description: Created user object
        required: true
  /user/login:
    get:
      tags:
        - user
      summary: Logs user into the system
      operationId: loginUser
      parameters: 
        - name: username
          in: query
          description: The user name for login
          required: true
          schema: 
            type: string
        - name: password
          in: query
          description: The password for login in clear text
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: string
            application/xml:
              schema:
                type: string
        '400':
          description: Invalid username/password supplied
...
  '/user/{username}':
    get:
      tags:
        - user
      summary: Get user by user name
      operationId: getUserByName
      parameters:
        - name: username
          in: path
          description: The name that needs to be fetched. Use user1 for testing.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/UserData'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
    put:
      tags:
        - user
      summary: Updated user
      description: This can only be done by the logged in user.
      operationId: updateUser
      parameters:
        - name: username
          in: path
          description: name that need to be updated
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid user supplied
        '404':
          description: User not found
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserData'
        description: Updated user object
        required: true
        
    delete:
      tags:
        - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
        - name: username
          in: path
          description: The name that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
...
  /fishingSpots/{name}:
    get:
      tags:
        - fishingSpots
      summary: Find fishing spots by name
      operationId: getFishingSpotsByName
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FishingSpots'          
            application/xml:
              schema:
                $ref: '#/components/schemas/FishingSpots'
        '400':
          description: Invalid name supplied
        '404':
          description: FishingSpots not found
      security:
        - fishingPortal_auth:
            - write:fishingSpots
            - read:fishingSpots
components:
  schemas:
    UserData:
      required:
        - id 
        - username
        - firstName
        - lastName
        - email
        - password
        - favoriteFishingSpots
      type: object
      properties:
        id: 
          type: integer
          format: int64
          example: 15
        username:
          type: string
          example: gunceOzer
        firstName: 
          type: string
          example: 'Gunce'
        lastName:
          type: string
          example: 'Ozer'
        email:
          type: string
          example: 'blabla@gmail.com'
        password:
          type: string
          example: '12345'
        favoriteFishingSpots:
          type: string
          example: 'Olesnica'
      xml:
        name: userData
    FishingSpots:
      required:
        - id
        - name
        - longtitude
        - altitude
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 3
        name: 
          type: string
          example: 'Olesnica'
        longtitude:
          type: number
          format: float
        altitude:
          type: number
          format: float
      xml:
        name: fishingSpots
...
  securitySchemes:
    fishingPortal_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://example.org/oauth/authorize
          scopes:
            write:fishingData: modify fishing data in your account
            read:fishingData: read fishing data